splitting pattern.cs file
