+++ FunctionalSharp/FunctionalSharp.PatternMatching/FunctionalSharp.PatternMatching.csproj
@@ -5,6 +5,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <None Include="Pattern{TResult}.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="FluentAssertions" Version="4.19.2" />
   </ItemGroup>
